+++ src/style.css
@@ -50,27 +50,20 @@ body {
 }
 
 
-/* Currently Playing takes up 3/4 of screen (right side) */
+/* Currently Playing in the center of the window */
 .currentlyPlaying {
-    flex: 3; /* 3/4 of screen */
+    flex: 1; /* 3/4 of screen */
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     padding: 2rem;
-    background-color: #1db954; /* Spotify green */
-    border-top-left-radius: 16px;
-    border-bottom-left-radius: 16px;
-    box-shadow: -4px 0 12px rgba(0,0,0,0.3);
+    background-color: transparent; /* Spotify green */
+    border-top-left-radius: 1px;
+    border-bottom-left-radius: 1px;
     text-align: center;
 }
 
-/* Left empty 1/4 area */
-body::before {
-    content: "";
-    flex: 1; /* 1/4 of screen */
-    background-color: #121212; /* dark left strip */
-}
 
 /* Song Title above album art */
 .currentlyPlaying h1 {
@@ -85,44 +78,86 @@ body::before {
     box-shadow: 0 4px 16px rgba(0,0,0,0.4);
 }
 
-
-/* Video and image fill the right 3/4 completely */
 #bgVideo,
 #bgImage {
     position: fixed;
-    top: 0;
-    left: 0;
+    top: 50%;
+    left: 50%;
     width: 100vw;
     height: 100vh;
-    object-fit: contain;       /* fit inside viewport, no cropping */
-    object-position: center;   /* center horizontally and vertically */
-    background-color: #000;    /* fallback for empty space */
+    transform: translate(-50%, -50%);
+    object-fit: cover;
+    object-position: center;
     z-index: 0;
 }
 
-#bgImage {
-    display: none; /* hidden by default, only shows if video fails */
-}
 
+#bgImage {
+    display: none; /* hidden by default */
+}
 
 /* Foreground content over video/image */
 .currentlyPlaying .content {
-    position: relative;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%); /* perfectly center */
     z-index: 2;
     text-align: center;
     color: white;
-    text-shadow: 0 2px 6px rgba(0,0,0,0.7);
+    text-shadow: 0 0 6px rgba(0,0,0,0.9), 0 2px 6px rgba(0,0,0,0.7); /* strong outline */
+    pointer-events: none;
+
+    /* Pulse animation (can be triggered on song change) */
+    animation: pulseBeat 1s ease-in-out infinite;
 }
+
+/* Song title styling */
+.currentlyPlaying .content h1 {
+    font-size: 4rem;
+    margin: 0;
+}
+
+/* Pulse animation keyframes */
+@keyframes pulseBeat {
+    0%, 100% {
+        transform: translate(-50%, -50%) scale(1);
+    }
+    50% {
+        transform: translate(-50%, -50%) scale(1.05);
+    }
+}
+
+
+/* Keyframes for floating effect */
+@keyframes floatTitle {
+    0%, 100% {
+        transform: translate(-50%, -50%) translateY(0px) scale(1);
+    }
+    50% {
+        transform: translate(-50%, -50%) translateY(-10px) scale(1.02);
+    }
+}
+
+/* Left column with up next songs */
 
 #upNextColumn {
     display: flex;
     flex-direction: column;
-    gap: 1rem;          /* spacing between songs */
+    gap: 1.15rem;        /* increased by 15% for balance */
     position: fixed;
-    left: 1rem;          /* move to left side */
+    left: 1rem;
     top: 20%;
-    width: 200px;         /* column width */
+    width: 200px;
 }
+
+.songItem img {
+    width: 172px;        /* 150px × 1.15 */
+    height: 172px;       /* 150px × 1.15 */
+    object-fit: cover;
+    border-radius: 10px;
+}
+
 
 .songItem {
     display: flex;
@@ -131,18 +166,15 @@ body::before {
     text-align: center;
 }
 
-.songItem img {
-    width: 150px;
-    height: 150px;
-    object-fit: cover;
-    border-radius: 10px;
-}
-
 .songItem span {
     margin-bottom: 0.5rem;
     font-weight: bold;
-    color: white;       /* adjust based on your background */
+    color: white;
+    text-shadow:
+            0 1px 2px rgba(0, 0, 0, 0.8),   /* subtle inner shadow */
+            0 2px 4px rgba(0, 0, 0, 0.7);   /* larger outer shadow for depth */
 }
+
